+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - MCP: Add subsystem providing a few server and client utilities through
   the `ctk query mcp {list,inquire,launch}` subcommands.
+- Docs API: Added CrateDB settings extractor
 
 ## 2025/01/31 v0.0.31
 - Fixed connectivity for `jobstats collect`